# test/swagger/oss_api.yaml
swagger: '2.0'
info:
  title: OSS管理系统API
  description: OSS管理系统后端API接口文档
  version: "1.0.0"
  contact:
    email: admin@example.com
host: localhost:8080
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: 在请求头中添加 "Bearer {token}" 进行身份验证
...
paths:
...
  /oss/files/{id}/md5:
    post:
      tags:
        - oss-files
      summary: 触发MD5计算
      operationId: triggerMD5Calculation
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: 文件ID
      responses:
        200:
          description: 触发成功
          schema:
            $ref: '#/definitions/Response'
        401:
          description: 未授权
          schema:
            $ref: '#/definitions/ErrorResponse'
        404:
          description: 文件不存在
          schema:
            $ref: '#/definitions/ErrorResponse'
        500:
          description: 服务器错误
          schema:
            $ref: '#/definitions/ErrorResponse'
    
    get:
      tags:
        - oss-files
      summary: 获取文件MD5值
      operationId: getFileMD5
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: 文件ID
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 0
              message:
                type: string
                example: 成功
              data:
                type: object
                properties:
                  md5:
                    type: string
                    description: MD5值
                    example: d41d8cd98f00b204e9800998ecf8427e
                  status:
                    type: string
                    description: 计算状态
                    enum: [pending, processing, completed, failed]
                    example: completed
        401:
          description: 未授权
          schema:
            $ref: '#/definitions/ErrorResponse'
        404:
          description: 文件不存在
          schema:
            $ref: '#/definitions/ErrorResponse'
        500:
          description: 服务器错误
          schema:
            $ref: '#/definitions/ErrorResponse'
...
  /oss/multipart/init:
    post:
      tags:
        - oss-multipart
      summary: 初始化分片上传
      operationId: initMultipartUpload
      security:
        - Bearer: []
      parameters:
        - in: query
          name: filename
          type: string
          required: true
          description: 文件名
          example: large_file.zip
        - in: query
          name: storage_type
          type: string
          required: false
          description: 存储类型，不传则使用默认存储服务
          enum: [ALIYUN_OSS, AWS_S3, CLOUDFLARE_R2]
      responses:
        200:
          description: 初始化成功
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 0
              message:
                type: string
                example: 初始化成功
              data:
                type: object
                properties:
                  upload_id:
                    type: string
                    description: 上传ID
                    example: abc123def456
                  urls:
                    type: array
                    description: 分片上传URL列表
                    items:
                      type: string
                      example: "https://example.com/upload/part1?token=abc"
        400:
          description: 请求参数错误
          schema:
            $ref: '#/definitions/ErrorResponse'
        401:
          description: 未授权
          schema:
            $ref: '#/definitions/ErrorResponse'
        500:
          description: 服务器错误
          schema:
            $ref: '#/definitions/ErrorResponse'
  /oss/multipart/complete:
    post:
      tags:
        - oss-multipart
      summary: 完成分片上传
      operationId: completeMultipartUpload
      security:
        - Bearer: []
      parameters:
        - in: body
          name: completeInfo
          description: 完成信息
          required: true
          schema:
            type: object
            required:
              - upload_id
              - filename
              - parts
            properties:
              upload_id:
                type: string
                description: 上传ID
                example: abc123def456
              filename:
                type: string
                description: 文件名
                example: large_file.zip
              parts:
                type: array
                description: 分片信息
                items:
                  type: object
                  properties:
                    PartNumber:
                      type: integer
                      description: 分片编号
                      example: 1
                    ETag:
                      type: string
                      description: ETag值
                      example: "\"etag1\""
      responses:
        200:
          description: 上传完成
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 0
              message:
                type: string
                example: 上传完成
              data:
                $ref: '#/definitions/OSSFile'
        400:
          description: 请求参数错误
          schema:
            $ref: '#/definitions/ErrorResponse'
        401:
          description: 未授权
          schema:
            $ref: '#/definitions/ErrorResponse'
        500:
          description: 服务器错误
          schema:
            $ref: '#/definitions/ErrorResponse'
  /oss/multipart/abort:
    delete:
      tags:
        - oss-multipart
      summary: 取消分片上传
      operationId: abortMultipartUpload
      security:
        - Bearer: []
      parameters:
        - in: query
          name: upload_id
          type: string
          required: true
          description: 上传ID
          example: abc123def456
        - in: query
          name: filename
          type: string
          required: true
          description: 文件名
          example: large_file.zip
        - in: query
          name: storage_type
          type: string
          required: false
          description: 存储类型，不传则使用默认存储服务
          enum: [ALIYUN_OSS, AWS_S3, CLOUDFLARE_R2]
      responses:
        200:
          description: 取消成功
          schema:
            $ref: '#/definitions/Response'
        400:
          description: 请求参数错误
          schema:
            $ref: '#/definitions/ErrorResponse'
        401:
          description: 未授权
          schema:
            $ref: '#/definitions/ErrorResponse'
        500:
          description: 服务器错误
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  Response:
    type: object
    properties:
      code:
        type: integer
        example: 0
      message:
        type: string
        example: 操作成功
      data:
        type: object
        description: 空对象或其他信息
  ErrorResponse:
    type: object
    properties:
      code:
        type: integer
        example: 1001
      message:
        type: string
        example: 参数错误
      data:
        type: object
        description: 错误详情
...
  OSSFile:
    type: object
    properties:
      id:
        type: integer
        example: 1
      file_name:
        type: string
        example: example.txt
      file_size:
        type: integer
        example: 1024
      file_type:
        type: string
        example: text/plain
      storage_type:
        type: string
        enum: [ALIYUN_OSS, AWS_S3, CLOUDFLARE_R2]
        example: AWS_S3
      object_key:
        type: string
        example: uploads/example.txt
      config_id:
        type: integer
        example: 1
      config_name:
        type: string
        example: AWS S3存储
      md5:
        type: string
        example: d41d8cd98f00b204e9800998ecf8427e
      download_url:
        type: string
        example: https://example.com/download/abc123
      created_at:
        type: string
        format: date-time
        example: "2023-01-01T12:00:00Z"
      updated_at:
        type: string
        format: date-time
        example: "2023-01-01T12:00:00Z"
...